[33] Add navbar and separate config component
